% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{car_cost}
\alias{car_cost}
\title{calculates the cost for a car, given a distance Units: $ / (Kg*Km)}
\usage{
car_cost(d)
}
\arguments{
\item{d}{the distance the car travels (in Km)}
}
\value{
the car cost for d distance
}
\description{
calculates the cost for a car, given a distance Units: $ / (Kg*Km)
}
\keyword{car}
\keyword{cost}
\keyword{vehichle}
