% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cost_per_kg_per_period}
\alias{cost_per_kg_per_period}
\title{calculates the cost per kg / period}
\usage{
cost_per_kg_per_period(d, t = 1)
}
\arguments{
\item{d}{the distance to the store}

\item{t}{the unit of time (number of periods)}
}
\value{
the truck cost for d distance
}
\description{
calculates the cost per kg / period
}
\keyword{cost}
\keyword{space}
\keyword{store}
