% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_costs_per_size}
\alias{get_costs_per_size}
\title{Returns the costs per size of the facility ($ / ft^2)}
\usage{
get_costs_per_size()
}
\value{
the cost per size of the facility
}
\description{
Returns the costs per size of the facility ($ / ft^2)
}
\keyword{cost}
\keyword{size}
\keyword{store}
