% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_demand}
\alias{get_demand}
\title{Returns the demand (kg of products a customer purchases in a period)}
\usage{
get_demand()
}
\value{
The demand
}
\description{
Returns the demand (kg of products a customer purchases in a period)
}
\keyword{demand}
