% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_floor_sizes}
\alias{get_floor_sizes}
\title{Returns the floor sizes of the facilities (ft^2)}
\usage{
get_floor_sizes()
}
\value{
the floor sizes of the facilities (ft^2)
}
\description{
Returns the floor sizes of the facilities (ft^2)
}
\keyword{floor}
\keyword{sizes}
\keyword{store}
