% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_store_capacities}
\alias{get_store_capacities}
\title{Returns the capacities of the facilities (number of people)}
\usage{
get_store_capacities()
}
\value{
the capacities of the facilities
}
\description{
Returns the capacities of the facilities (number of people)
}
\keyword{capactity}
\keyword{store}
