\name{gsc}
\alias{gsc}
\title{generalised shape constraints}
\description{This function creates a gsc object, to represent generalised shape constraints.}
\usage{
gsc (k=numeric (), d0=NA, d1=NA, d2=NA)
}
\arguments{
\item{k}{Breakpoints for a function's domain.}
\item{d0}{Constraint vector for 0th derivative.}
\item{d1}{Constraint vector for 1th derivative.}
\item{d2}{Constraint vector for 2th derivative.}
}

