% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{locate_warehouse}
\alias{locate_warehouse}
\title{Centrally locates a warehouse across a region defined by the data frame x}
\usage{
locate_warehouse(x)
}
\arguments{
\item{x}{the data frame containing locations to locate the warehouse across}
}
\description{
Centrally locates a warehouse across a region defined by the data frame x
}
