% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{space_cost}
\alias{space_cost}
\title{calculates the space cost for a facility. Units: $}
\usage{
space_cost(v, f, p_f, e, p_e, q)
}
\arguments{
\item{v}{variable space cost (eg rent) / time}

\item{f}{amount of energy / time}

\item{p_f}{price of energy ($ / unit of energy)}

\item{e}{amount of emissions / unit of energy}

\item{p_e}{price of emissions}
}
\value{
the truck cost for d distance
}
\description{
calculates the space cost for a facility. Units: $
}
\keyword{cost}
\keyword{store}
