% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{store_cost_foot}
\alias{store_cost_foot}
\title{calculates the space cost for a facility. Units: $ / sqft}
\usage{
store_cost_foot(sq_ft)
}
\value{
the truck cost for d distance
}
\description{
calculates the space cost for a facility. Units: $ / sqft
}
\keyword{cost}
\keyword{space}
\keyword{store}
