% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{truck_cost}
\alias{truck_cost}
\title{calculates the cost for a truck, given a distance. Units: $ / (Kg*Km)}
\usage{
truck_cost(d)
}
\arguments{
\item{d}{the distance the truck travels (in Km)}
}
\value{
the truck cost for d distance
}
\description{
calculates the cost for a truck, given a distance. Units: $ / (Kg*Km)
}
\keyword{cost}
\keyword{truck}
\keyword{vehichle}
