% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vehicle_cost}
\alias{vehicle_cost}
\title{calculates the cost for a vehichle. Units: $ / (Kg*Km)}
\usage{
vehicle_cost(v, f, p_f, e, p_e, q)
}
\arguments{
\item{v}{nonfuel variable cost / distance}

\item{f}{amount of fuel / distance}

\item{p_f}{price of fuel ($ / amount of fuel)}

\item{e}{amount of emissions / unit of fuel}

\item{p_e}{price of emissions}

\item{q}{load carried by the vehicle}
}
\value{
the vehicle cost
}
\description{
calculates the cost for a vehichle. Units: $ / (Kg*Km)
}
\keyword{car}
\keyword{cost}
\keyword{truck}
\keyword{vehichle}
